requirements.txt
Django==4.2.7
django-jazzmin==2.6.0
psycopg==3.1.14
django-filter==23.5
python-dotenv==1.0.0
django-solo==2.1.0
django-crum==0.7.9
django-extensions==3.2.3
psycopg2-binary==2.9.9; sys_platform == 'win32'
Faker==20.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-jazzmin==2.6.0
python-dotenv==1.0.0
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
django-reportbroD==3.0



